Move getSelectedTime() and getSelectedDate() to dateTimeFieldHelpers.js and import in NewTimeLogDateTimeFields and EditTimeLogDateTimeFields
