+++ src/components/WorkOrder/EditTimeLogDateTimeFields.js
@@ -2,7 +2,6 @@ import React from "react";
 import TimeLogForm from "./TimeLogForm";
 import { FIELDS } from "./formConfig";
 import moment from "moment";
-import { isEmpty } from "lodash";
 
 import {
   getHours,
@@ -10,6 +9,8 @@ import {
   convertKnackDateTimeToFormDate,
   convertKnackDateTimeToFormTime,
   addMissingFieldsWithExistingKnackData,
+  getSelectedDate,
+  getSelectedTime,
 } from "../Shared/dateTimeFieldHelpers.js";
 
 const EditTimeLogDateTimeFields = ({
@@ -53,25 +54,14 @@ const EditTimeLogDateTimeFields = ({
 
   const getFormSelection = (field, dateOrTime) => {
     if (dateOrTime === "time") {
-      return getExistingRecord(field, dateOrTime) || getSelectedTime(field);
+      return (
+        getExistingRecord(field, dateOrTime) || getSelectedTime(data, field)
+      );
     } else if (dateOrTime === "date") {
-      return getExistingRecord(field, dateOrTime) || getSelectedDate(field);
+      return (
+        getExistingRecord(field, dateOrTime) || getSelectedDate(data, field)
+      );
     }
-  };
-
-  const getSelectedDate = field => {
-    // When the form first loads, autofill current date & time
-    // on certain fields to mirror Knack data validation.
-    const today = new Date();
-    const shouldReturnCurrentDate =
-      (isEmpty(data[field]) && field === FIELDS.TIMELOG.ISSUE_RECEIVED_TIME) ||
-      (isEmpty(data[field]) && field === FIELDS.TIMELOG.WORKSITE_ARRIVE);
-
-    if (shouldReturnCurrentDate) return today;
-    if (isEmpty(data[field])) return null;
-
-    let date = data[field].date ? new Date(data[field].date) : today;
-    return date;
   };
 
   // Get updated form data to populate picker field (state after update from Edit Form)
@@ -90,28 +80,6 @@ const EditTimeLogDateTimeFields = ({
     } else {
       return false;
     }
-  };
-
-  const getSelectedTime = field => {
-    // When the form first loads, autofill current date & time
-    // on certain fields to mirror Knack data validation.
-    const today = new Date();
-    const shouldReturnCurrentDate =
-      (isEmpty(data[field]) && field === FIELDS.TIMELOG.ISSUE_RECEIVED_TIME) ||
-      (isEmpty(data[field]) && field === FIELDS.TIMELOG.WORKSITE_ARRIVE);
-
-    if (shouldReturnCurrentDate) return today; // default to current time for certain fields
-    if (isEmpty(data[field])) return null; // let value start as blank if there's no exisiting data
-
-    let date = data[field].date
-      ? data[field].date
-      : moment(Date.now()).format("MM/DD/YYYY");
-
-    let momentDate = moment(
-      `${date} ${data[field].hours}:${data[field].minutes} ${data[field].am_pm}`
-    ).format("MM/DD/YYYY h:mm a");
-
-    return new Date(momentDate);
   };
 
   const updateErrorState = () => {
